refactor: extract helper functions and better comment
